refactor(rolldown_plugin_build_import_analysis): extract AST visit logic
